# internal/rest/v1/boilerplate/server.gen.go
// Package boilerplate provides primitives to interact with the openapi HTTP API.
//
// Code generated by github.com/oapi-codegen/oapi-codegen/v2 version v2.4.1 DO NOT EDIT.
package boilerplate

import (
	"fmt"
	"net/http"

	"github.com/labstack/echo/v4"
	"github.com/oapi-codegen/runtime"
)

// ServerInterface represents all server handlers.
type ServerInterface interface {
	// serve index.html
	// (GET /)
	ServeIndex(ctx echo.Context) error
	// healthcheck endpoint
	// (GET /healthcheck)
	Healthcheck(ctx echo.Context) error
	// create new room
	// (POST /room)
	CreateNewRoom(ctx echo.Context) error
	// get room info
	// (GET /room/{id})
	GetRoomInfo(ctx echo.Context, id string) error
	// returns json api description
	// (GET /swagger)
	GetApi(ctx echo.Context) error
	// connect to room
	// (GET /ws/room/{id})
	ConnectRoom(ctx echo.Context, id string, params ConnectRoomParams) error
}

// ServerInterfaceWrapper converts echo contexts to parameters.
type ServerInterfaceWrapper struct {
	Handler ServerInterface
}

// ServeIndex converts echo context to params.
func (w *ServerInterfaceWrapper) ServeIndex(ctx echo.Context) error {
	var err error

	// Invoke the callback with all the unmarshaled arguments
	err = w.Handler.ServeIndex(ctx)
	return err
}

// Healthcheck converts echo context to params.
func (w *ServerInterfaceWrapper) Healthcheck(ctx echo.Context) error {
	var err error

	// Invoke the callback with all the unmarshaled arguments
	err = w.Handler.Healthcheck(ctx)
	return err
}

// CreateNewRoom converts echo context to params.
func (w *ServerInterfaceWrapper) CreateNewRoom(ctx echo.Context) error {
	var err error

	// Invoke the callback with all the unmarshaled arguments
	err = w.Handler.CreateNewRoom(ctx)
	return err
}

// GetRoomInfo converts echo context to params.
func (w *ServerInterfaceWrapper) GetRoomInfo(ctx echo.Context) error {
	var err error
	// ------------- Path parameter "id" -------------
	var id string

	err = runtime.BindStyledParameterWithOptions("simple", "id", ctx.Param("id"), &id, runtime.BindStyledParameterOptions{ParamLocation: runtime.ParamLocationPath, Explode: false, Required: true})
	if err != nil {
		return echo.NewHTTPError(http.StatusBadRequest, fmt.Sprintf("Invalid format for parameter id: %s", err))
	}

	// Invoke the callback with all the unmarshaled arguments
	err = w.Handler.GetRoomInfo(ctx, id)
	return err
}

// GetApi converts echo context to params.
func (w *ServerInterfaceWrapper) GetApi(ctx echo.Context) error {
	var err error

	// Invoke the callback with all the unmarshaled arguments
	err = w.Handler.GetApi(ctx)
	return err
}

// ConnectRoom converts echo context to params.
func (w *ServerInterfaceWrapper) ConnectRoom(ctx echo.Context) error {
	var err error
	// ------------- Path parameter "id" -------------
	var id string

	err = runtime.BindStyledParameterWithOptions("simple", "id", ctx.Param("id"), &id, runtime.BindStyledParameterOptions{ParamLocation: runtime.ParamLocationPath, Explode: false, Required: true})
	if err != nil {
		return echo.NewHTTPError(http.StatusBadRequest, fmt.Sprintf("Invalid format for parameter id: %s", err))
	}

	// Parameter object where we will unmarshal all parameters from the context
	var params ConnectRoomParams
	// ------------- Optional query parameter "username" -------------

	err = runtime.BindQueryParameter("form", true, false, "username", ctx.QueryParams(), &params.Username)
	if err != nil {
		return echo.NewHTTPError(http.StatusBadRequest, fmt.Sprintf("Invalid format for parameter username: %s", err))
	}

	// Invoke the callback with all the unmarshaled arguments
	err = w.Handler.ConnectRoom(ctx, id, params)
	return err
}

// This is a simple interface which specifies echo.Route addition functions which
// are present on both echo.Echo and echo.Group, since we want to allow using
// either of them for path registration
type EchoRouter interface {
	CONNECT(path string, h echo.HandlerFunc, m ...echo.MiddlewareFunc) *echo.Route
	DELETE(path string, h echo.HandlerFunc, m ...echo.MiddlewareFunc) *echo.Route
	GET(path string, h echo.HandlerFunc, m ...echo.MiddlewareFunc) *echo.Route
	HEAD(path string, h echo.HandlerFunc, m ...echo.MiddlewareFunc) *echo.Route
	OPTIONS(path string, h echo.HandlerFunc, m ...echo.MiddlewareFunc) *echo.Route
	PATCH(path string, h echo.HandlerFunc, m ...echo.MiddlewareFunc) *echo.Route
	POST(path string, h echo.HandlerFunc, m ...echo.MiddlewareFunc) *echo.Route
	PUT(path string, h echo.HandlerFunc, m ...echo.MiddlewareFunc) *echo.Route
	TRACE(path string, h echo.HandlerFunc, m ...echo.MiddlewareFunc) *echo.Route
}

// RegisterHandlers adds each server route to the EchoRouter.
func RegisterHandlers(router EchoRouter, si ServerInterface) {
	RegisterHandlersWithBaseURL(router, si, "")
}

// Registers handlers, and prepends BaseURL to the paths, so that the paths
// can be served under a prefix.
func RegisterHandlersWithBaseURL(router EchoRouter, si ServerInterface, baseURL string) {

	wrapper := ServerInterfaceWrapper{
		Handler: si,
	}

	router.GET(baseURL+"/", wrapper.ServeIndex)
	router.GET(baseURL+"/healthcheck", wrapper.Healthcheck)
	router.POST(baseURL+"/room", wrapper.CreateNewRoom)
	router.GET(baseURL+"/room/:id", wrapper.GetRoomInfo)
	router.GET(baseURL+"/swagger", wrapper.GetApi)
	router.GET(baseURL+"/ws/room/:id", wrapper.ConnectRoom)

}